+++ src/descriptors/_element_edge_test.js
@@ -67,7 +67,9 @@ describe("ElementEdge", function() {
 		}
 	});
 
-	it.skip("accounts for scrolling", function() {
+	it("accounts for scrolling", function() {
+		if (!quixote.browser.canScroll()) return;
+
 		frame.addElement("<div style='position: absolute; left: 5000px; top: 5000px; width: 60px'>scroll enabler</div>");
 
 		frame.scroll(50, 60);